requirements.txt
runpod
torch
diffusers
transformers
safetensors
scikit-learn
pillow
